Add an option to deactivate input assistant integration
Our plugin changes fields order for the address, deactivates autocomplete and host specific field modifications in order for
our autcomplete to work. Sometimes and in some countries like USA it might not be desired, as our approach to input assistant it
foreign to the users in those countries.

Because we currently don't have a way to provide different strategies for autocomplete depending on the country, we should at least
provide an option to deactivate the autocomplete and its preparation routines.

Which is done in this commit. The input assistant is now a separate function that can be deactivated. Address check still works, but all field
manipulations and integration of autocomplete are skipped, the browser autocomplete works.