Fix subdivision initial read, email and address validation issues
In the QA of RC1 we found a few problems. Some of them come from JS-SDK and a fix will be provided in 1.12.1 (this commit will then be amended).
Some are caused by ACCESSIBILITY_TWEAKS set true. Its changing the way templates work in shopware which breaks email check and subdivision initial
read. We also noticed, that additional info is not dsplayed in fake form of checkout last page, which break the address check.

This commit contains fixes for those problems and is a foundation for RC.2

Ref: S6A-272