Integrate js-sdk 1.13.0
In 1.13.0 js-sdk provides some new functionality and bug fixes we need.
One of the bug fixes is a fix with multiple levels of address check, where
an address check on level 0 can trigger an address check on level 1, which
blocks all address checks on level 0. This case wasn't handled properly before
and the update to 1.13.0 provides a solution.
In endereco.js we need to provide domain knowledge on when the level 1 is
reached, which is when the address modal from shopware is open.

Ref: DEV-23